requirements.txt
azure-core==1.30.1
azure-cosmos==4.7.0
azure-messaging-webpubsubservice==1.1.0
azure-servicebus==7.12.2
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
